requirements.txt
numpy==1.21.6
pandas==1.3.5
Pillow==9.4.0
python-dateutil==2.8.2
pytz==2022.7
six==1.16.0